requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.6.0
wildlife-datasets>=0.1.0
wildlife-tools>=0.1.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
Pillow>=8.3.0
tqdm>=4.62.0
PyYAML>=5.4.0
matplotlib>=3.4.0
seaborn>=0.11.0
rawpy>=0.16.0